Remove chat components and add GameDetailMenu for improved game detail navigation
