+++ src/screens/game-detail/index.ts
@@ -1,10 +1,8 @@
 export * from "./channel-list";
 export * from "./channel";
-export * from "./chat";
 export * from "./create-channel";
 export * from "./map";
 export * from "./orders";
 export * from "./game-detail-layout";
-export * from "./chat-layout";
 export * from "./game-detail";
 export * from "./map-orders-layout"
